rdma: accelerated memcpy() support and better external RDMA user interfaces
This patch implements the ability to provide accelerated
memory copies as an alternative to memcpy() using RDMA.

It requires the user to "register" the src and dest memory
with the RDMA subsystem using new interfaces and then call
a special copy function that is introduced to submit the
copy to the hardware.

This also significantly re-works the RDMA code itself such that
it is more agile: Users can call into the RDMA code from arbitrary
parts of the QEMU code base to create new connections or send
memory across an existing connection. (While this has not yet been
tested, with say, storage or something similar, the hardwork of
fixing the RDMA plumbing to make it work has mostly been done).