Believe it or not, when there is no network device attached to the guest whatsoever, the intial bootup process can be extremely slow, where there are almost no processes dirtying memory at all or only occasionally except for maybe a DHCP client. This results in some 100ms periods of time where there are actually *no* dirty pages.
ram_save_block() really doesn't understand this possibility,
surprisingly. It results in an infinite loop because it was expecting
last_seen_block to always be non-NULL, when in fact, we have reset
the value to start from the beginning of the guest can scan the
entire VM for dirty memory.